pyproject.toml
[project]
name = "crypto-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "annotated-doc==0.0.4",
    "celery[redis]>=5.6.3",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "django>=6.0.3",
    "django-debug-toolbar>=6.3.0",
    "django-environ>=0.13.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "idna==3.11",
    "ipython>=9.12.0",
    "librt==0.8.1",
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    "pathspec==1.0.4",
    "psycopg2-binary>=2.9.11",
    "pygments==2.19.2",
    "python-dotenv==1.2.2",
    "requests==2.32.5",
    "rich==14.3.3",
    "shellingham==1.5.4",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "tzdata>=2025.3",
    "urllib3==2.6.3",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=6.0.3",
    "djangorestframework-stubs[compatible-mypy]>=3.16.9",
    "mypy>=1.19.1",
    "pre-commit>=4.6.0",
    "ruff>=0.15.11",
    "types-requests>=2.33.0.20260408",
    { include-group = "test" },
]

test = [
    "freezegun>=1.5.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-django>=4.12.0",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
markers = [
    "unit: Quick tests with mocks (no I/O or DB)",
    "integration: Tests that interact with DB or File System",
]
addopts = [
    "-v",
    "-ra",
    "--tb=short",
    "--strict-markers",
    "--cov=src",
    "--cov=crypto",
    "--cov-report=term-missing",
]
DJANGO_SETTINGS_MODULE = "config.settings"

[tool.coverage.report]
exclude_lines = [
    # Игнорировать стандартную проверку запуска модуля
    "if __name__ == .__main__.:",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
check_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true

[tool.django-stubs]
django_settings_module = "config.settings"

[[tool.mypy.overrides]]
module = ["environ", "environ.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["celery", "celery.*"]
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py314"
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 72
